# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for short url.",
        "title": "ShortURL API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/shorten": {
            "post": {
                "description": "Create a short url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortURL"
                ],
                "summary": "Create a short url",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.URLShortenCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.URLShortenCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stats/{shortCode}": {
            "get": {
                "description": "Returns information such as clicks and createdAt",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShortURL"
                ],
                "summary": "Get statistics for a short URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short Code",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.URLShortenGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/{shortCode}": {
            "get": {
                "description": "Redirect user using short code",
                "tags": [
                    "ShortURL"
                ],
                "summary": "Redirect to the original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short Code",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.URLShortenCreateRequest": {
            "type": "object",
            "required": [
                "originalUrl"
            ],
            "properties": {
                "originalUrl": {
                    "type": "string"
                },
                "shortCode": {
                    "type": "string",
                    "default": "abc123"
                }
            }
        },
        "entities.URLShortenCreateResponse": {
            "type": "object",
            "properties": {
                "originalUrl": {
                    "type": "string"
                },
                "shortCode": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                }
            }
        },
        "entities.URLShortenGetResponse": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "originalUrl": {
                    "type": "string"
                },
                "shortCode": {
                    "type": "string"
                }
            }
        }
    }
}